Move base into v04
