+++ src/ome_zarr_models/v04/base.py
