+++ src/ome_zarr_models/v04/plate.py
@@ -14,7 +14,7 @@ from pydantic import (
 )
 
 from ome_zarr_models._utils import _AlphaNumericConstraint, _unique_items_validator
-from ome_zarr_models.base import Base
+from ome_zarr_models.v04.base import Base
 
 __all__ = [
     "Acquisition",
